requirements.txt
numpy>=1.15.0
python-dateutil>=2.7.3
requests>=2.19.1
